Fix elm-css cli for Elm version 0.17.
Due to the changes in the port system on both the JavaScript and Elm
side the cli application was no longer able to extract the generated
CSS from the Elm appliation.

This commit fixes the issue in the CLI and updates the example code so
the two work together again.

Resolves #120.