Address review remarks
Replaced `pointerEventsVisible` with `pointerEvents visible` — changed
the definition of `visible` to fit by following the example of `scroll`.
If that doesn’t work well, I’m open to ideas and guidance.

As for the `pointerEvents fill` and `pointerEvents all`, it’s my
understanding that we can’t have `fill` and `all` **as values** because
they already exist to designate **properties**. :neckbeard: So besides
`pointerEventsFill` I have also added `pointerEventsAll`.

Added descriptions and links for the values, and also listed the new
things in the Properties, Other values, and Misc `@docs` sections.