Remove logo.
I'm increasingly uncomfortable about the idea of Elm packages having logos.
I'd rather see an Elm packages distinguish itself through good design
and documentation in its modules than good visual design in its README.

I think this logo looks great, which is kind of the problem. I don't
want peoples' first impression to be the subconscious cue that the package
is good because the logo is good. I'd rather their first impression be the
words in the README itself. If I've done a good job explaining what the
package is for, and how to use it, that's at least a genuine signal that the
docs are likely to be worth reading.

The cultural value of having logos on packages has not served the JS ecosystem
well in my opinion, and I don't think it would serve the Elm ecosystem
well either.